Address static analysis reports
Use null propagation
Private members that can or should be readonly
Reduce allocations, garbage, and enumerations with this List
Add TryGetArray<T>
Using framework naming convention for methods that return boolean status and have out parameters that are non-null on success
There's another use of the old method, but I have to stare at it a bit more to see what it's really doing.
Might need to discuss that one when we chat.