Add TryGetArray<T>
Replaces GetArray
Using framework naming convention for methods that return boolean status and have out parameters that are non-null on success
There's another use of the old method, but I have to stare at it a bit more to see what it's really doing.
Might need to discuss that one when we chat.